user components editted, transaction components started

+++ src/components/User/UserDetail.js
@@ -8,7 +8,7 @@ const User = (props) => {
 	const [error, setError] = useState(false);
 
 	useEffect(() => {
-		const url = `${APIURL}/api/user/`;
+		const url = `${APIURL}/api/user`;
 		fetch(url, {
 			method: 'GET',
 			headers: {
@@ -24,34 +24,60 @@ const User = (props) => {
 	}, []);
 
 	const onDeleteUser = (event) => {
-		const url = `${APIURL}/api/user/`;
-		fetch(url, {
-			method: 'DELETE',
-			headers: {
-				'Content-Type': 'application/json',
-				Accept: 'application/json',
-				Authorization: `Bearer ${props.userToken}`,
-			},
-		})
-			.then((res) => {
-				setDeleted(true);
+		let confirm = prompt(
+			"This action will delete the current user. Please type 'confirm' to delete", ''
+		);
+		if (confirm === 'confirm') {
+			const url = `${APIURL}/api/user`;
+			fetch(url, {
+				method: 'DELETE',
+				headers: {
+					'Content-Type': 'application/json',
+					Accept: 'application/json',
+					Authorization: `Bearer ${props.userToken}`,
+				},
 			})
-			.catch(console.error);
+				.then((res) => {
+					setDeleted(true);
+				})
+				.catch(console.error);
+		}
 	};
 	if (deleted) {
 		props.handleSignOut();
-		return <Redirect to='/' />;
+		return <Redirect to='/signin' />;
 	}
 
-    return (
+	return (
 		<div className='user-account-wrapper'>
 			{!user ? (
-				<div className='home-title'>Sorry, there was a problem getting the users</div>
-			) : (
-				<div className='user-account-info'>
-					<h1>User Account</h1>
+				<div className='home-title'>
+					Sorry, there was a problem getting the users
 				</div>
-			)}{' '}
+			) : (
+				<>
+					<div className='user-account-info'>
+						<h1>Your Account</h1>
+						<h2>{user.firstname}</h2>
+						<h2>{user.lastname}</h2>
+						<h2>{user.username}</h2>
+						<h2>{user.email}</h2>
+
+					</div>
+					<Link to='/user/new-transaction'>Create New Transaction</Link>
+					<Link to='/user/all-transactions'>View Your Transaction History</Link>
+					<Link to='/user/balance-statement'>View Your Balance Statement</Link>
+
+					<div className='mt-5 link'>
+						<Link className='btn btn-info item' to='/user/edit'>
+							Update Account Information
+						</Link>
+					</div>
+					<button onClick={onDeleteUser} className='btn btn-danger item'>
+						Delete Account
+					</button>
+				</>
+			)}
 		</div>
 	);
 };
